requirements.txt
astroid==2.5.2
cxxfilt==0.2.2
isort==5.8.0
jedi==0.18.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.20.2
nvidia-pyprof==3.11.0
parso==0.8.2
pep8==1.7.1
Pillow==8.2.0
pylint==2.7.4
toml==0.10.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.60.0
typing-extensions==3.7.4.3
wrapt==1.12.1
yapf==0.31.0
